# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/enterprise/structure": {
      "get": {
        "operationId": "EnterpriseStructureController_getStructure",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "All three dimensions side by side, plus what is unplaced in each.",
        "tags": [
          "EnterpriseStructure"
        ]
      }
    },
    "/api/v1/enterprise/groups": {
      "post": {
        "operationId": "EnterpriseStructureController_createGroup",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGroupDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "EnterpriseStructure"
        ]
      },
      "get": {
        "operationId": "EnterpriseStructureController_listGroups",
        "parameters": [
          {
            "name": "includeInactive",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 120,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "EnterpriseStructure"
        ]
      }
    },
    "/api/v1/enterprise/groups/{id}": {
      "get": {
        "operationId": "EnterpriseStructureController_getGroup",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "EnterpriseStructure"
        ]
      },
      "patch": {
        "operationId": "EnterpriseStructureController_updateGroup",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateGroupDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "EnterpriseStructure"
        ]
      },
      "delete": {
        "operationId": "EnterpriseStructureController_deleteGroup",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Refused where history or sites exist — deactivate those instead.",
        "tags": [
          "EnterpriseStructure"
        ]
      }
    },
    "/api/v1/enterprise/groups/{id}/deactivate": {
      "post": {
        "operationId": "EnterpriseStructureController_deactivateGroup",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Retires a group without touching the sites in it.",
        "tags": [
          "EnterpriseStructure"
        ]
      }
    },
    "/api/v1/enterprise/groups/{id}/reactivate": {
      "post": {
        "operationId": "EnterpriseStructureController_reactivateGroup",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "EnterpriseStructure"
        ]
      }
    },
    "/api/v1/enterprise/groups/{id}/sites": {
      "post": {
        "operationId": "EnterpriseStructureController_assignSitesToGroup",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignSitesDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "EnterpriseStructure"
        ]
      }
    },
    "/api/v1/enterprise/groups/sites/{siteId}": {
      "delete": {
        "operationId": "EnterpriseStructureController_unassignSiteFromGroup",
        "parameters": [
          {
            "name": "siteId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "EnterpriseStructure"
        ]
      }
    },
    "/api/v1/enterprise/clusters": {
      "post": {
        "operationId": "EnterpriseStructureController_createCluster",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateClusterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "EnterpriseStructure"
        ]
      },
      "get": {
        "operationId": "EnterpriseStructureController_listClusters",
        "parameters": [
          {
            "name": "includeInactive",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 120,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "EnterpriseStructure"
        ]
      }
    },
    "/api/v1/enterprise/clusters/{id}": {
      "get": {
        "operationId": "EnterpriseStructureController_getCluster",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "EnterpriseStructure"
        ]
      },
      "patch": {
        "operationId": "EnterpriseStructureController_updateCluster",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateClusterDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "EnterpriseStructure"
        ]
      },
      "delete": {
        "operationId": "EnterpriseStructureController_deleteCluster",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "EnterpriseStructure"
        ]
      }
    },
    "/api/v1/enterprise/clusters/{id}/deactivate": {
      "post": {
        "operationId": "EnterpriseStructureController_deactivateCluster",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "EnterpriseStructure"
        ]
      }
    },
    "/api/v1/enterprise/clusters/{id}/reactivate": {
      "post": {
        "operationId": "EnterpriseStructureController_reactivateCluster",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "EnterpriseStructure"
        ]
      }
    },
    "/api/v1/enterprise/clusters/{id}/sites": {
      "post": {
        "operationId": "EnterpriseStructureController_assignSitesToCluster",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignSitesDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "EnterpriseStructure"
        ]
      }
    },
    "/api/v1/enterprise/clusters/sites/{siteId}": {
      "delete": {
        "operationId": "EnterpriseStructureController_unassignSiteFromCluster",
        "parameters": [
          {
            "name": "siteId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "EnterpriseStructure"
        ]
      }
    },
    "/api/v1/enterprise/territories": {
      "post": {
        "operationId": "EnterpriseStructureController_createTerritory",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTerritoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "EnterpriseStructure"
        ]
      },
      "get": {
        "operationId": "EnterpriseStructureController_listTerritories",
        "parameters": [
          {
            "name": "includeInactive",
            "required": false,
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "maxLength": 120,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "EnterpriseStructure"
        ]
      }
    },
    "/api/v1/enterprise/territories/{id}": {
      "get": {
        "operationId": "EnterpriseStructureController_getTerritory",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "EnterpriseStructure"
        ]
      },
      "patch": {
        "operationId": "EnterpriseStructureController_updateTerritory",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTerritoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "EnterpriseStructure"
        ]
      },
      "delete": {
        "operationId": "EnterpriseStructureController_deleteTerritory",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "EnterpriseStructure"
        ]
      }
    },
    "/api/v1/enterprise/territories/{id}/deactivate": {
      "post": {
        "operationId": "EnterpriseStructureController_deactivateTerritory",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "EnterpriseStructure"
        ]
      }
    },
    "/api/v1/enterprise/territories/{id}/reactivate": {
      "post": {
        "operationId": "EnterpriseStructureController_reactivateTerritory",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "EnterpriseStructure"
        ]
      }
    },
    "/api/v1/enterprise/territories/{id}/sites": {
      "post": {
        "operationId": "EnterpriseStructureController_assignSitesToTerritory",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignSitesDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "EnterpriseStructure"
        ]
      }
    },
    "/api/v1/enterprise/territories/sites/{siteId}": {
      "delete": {
        "operationId": "EnterpriseStructureController_unassignSiteFromTerritory",
        "parameters": [
          {
            "name": "siteId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "EnterpriseStructure"
        ]
      }
    },
...
    "/api/v1/admin/enterprise/{organisationId}/structure": {
      "get": {
        "operationId": "EnterpriseProvisioningController_getStructure",
        "parameters": [
          {
            "name": "organisationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "EnterpriseProvisioning"
        ]
      }
    },
...
  },
  "info": {
    "title": "Saveful for Business API",
    "description": "Food rescue platform API. All routes are prefixed `/api/v1`.\n\n**Authentication.** Every endpoint requires a bearer token unless marked otherwise. Obtain one from `POST /auth/login`, then send it as `Authorization: Bearer <token>`.\n\n**Enterprise endpoints** additionally require the organisation to be on the Enterprise plan, and are gated on the caller’s Enterprise role and scope. A 403 carrying `MISSING_PERMISSION` means the role is wrong; one carrying `OUTSIDE_SCOPE` means the role is right but the target sits outside what the caller can reach.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateGroupDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120
          },
          "code": {
            "type": "string",
            "maxLength": 40
          },
          "description": {
            "type": "string",
            "maxLength": 500
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateGroupDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120
          },
          "code": {
            "type": "string",
            "maxLength": 40
          },
          "description": {
            "type": "string",
            "maxLength": 500
          }
        }
      },
      "AssignSitesDto": {
        "type": "object",
        "properties": {
          "siteIds": {
            "minItems": 1,
            "type": "array",
            "items": {
              "type": "number"
            }
          }
        },
        "required": [
          "siteIds"
        ]
      },
      "CreateClusterDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120
          },
          "code": {
            "type": "string",
            "maxLength": 40
          },
          "description": {
            "type": "string",
            "maxLength": 500
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateClusterDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120
          },
          "code": {
            "type": "string",
            "maxLength": 40
          },
          "description": {
            "type": "string",
            "maxLength": 500
          }
        }
      },
      "CreateTerritoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120
          },
          "code": {
            "type": "string",
            "maxLength": 40
          },
          "description": {
            "type": "string",
            "maxLength": 500
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateTerritoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120
          },
          "code": {
            "type": "string",
            "maxLength": 40
          },
          "description": {
            "type": "string",
            "maxLength": 500
          }
        }
      },
...
    }
  }
}